[IT-2407] Renamed the wptserverproxy package to wptserver
