+++ grails-app/domain/de/iteratec/osm/report/external/GraphiteServer.groovy
@@ -17,9 +17,8 @@
 
 package de.iteratec.osm.report.external
 
-import de.iteratec.osm.measurement.environment.wptserverproxy.Protocol
-import de.iteratec.osm.report.external.provider.*
-import grails.databinding.BindUsing
+import de.iteratec.osm.measurement.environment.wptserver.Protocol
+import de.iteratec.osm.report.external.provider.GraphiteSocketProvider
 import grails.gorm.annotation.Entity
 
 /**
