+++ src/integration-test/groovy/de/iteratec/osm/csi/WeeklyPageIntegrationSpec.groovy
@@ -20,18 +20,14 @@ package de.iteratec.osm.csi
 import de.iteratec.osm.OsmConfiguration
 import de.iteratec.osm.measurement.environment.Browser
 import de.iteratec.osm.measurement.environment.Location
-import de.iteratec.osm.measurement.environment.wptserverproxy.ResultPersisterService
+import de.iteratec.osm.measurement.environment.wptserver.ResultPersisterService
 import de.iteratec.osm.measurement.schedule.ConnectivityProfile
 import de.iteratec.osm.measurement.schedule.Job
 import de.iteratec.osm.measurement.schedule.JobGroup
 import de.iteratec.osm.report.chart.AggregationType
 import de.iteratec.osm.report.chart.CsiAggregation
 import de.iteratec.osm.report.chart.CsiAggregationInterval
-import de.iteratec.osm.result.CachedView
-import de.iteratec.osm.result.JobResult
-import de.iteratec.osm.result.MeasuredEvent
-import de.iteratec.osm.result.WptStatus
-import de.iteratec.osm.result.EventResult
+import de.iteratec.osm.result.*
 import grails.gorm.transactions.Rollback
 import grails.testing.mixin.integration.Integration
 import org.joda.time.DateTime
