+++ src/integration-test/groovy/de/iteratec/osm/csi/WeeklyShopMultipleCsiGroupsIntegrationSpec.groovy
@@ -19,7 +19,7 @@ package de.iteratec.osm.csi
 
 import de.iteratec.osm.OsmConfiguration
 import de.iteratec.osm.measurement.environment.Browser
-import de.iteratec.osm.measurement.environment.wptserverproxy.ResultPersisterService
+import de.iteratec.osm.measurement.environment.wptserver.ResultPersisterService
 import de.iteratec.osm.measurement.schedule.ConnectivityProfile
 import de.iteratec.osm.measurement.schedule.Job
 import de.iteratec.osm.measurement.schedule.JobGroup
