+++ src/integration-test/groovy/de/iteratec/osm/measurement/environment/wptserver/PersistingResultsIntegrationSpec.groovy
@@ -15,7 +15,7 @@
 * limitations under the License.
 */
 
-package de.iteratec.osm.measurement.environment.wptserverproxy
+package de.iteratec.osm.measurement.environment.wptserver
 
 import de.iteratec.osm.OsmConfiguration
 import de.iteratec.osm.csi.CsiAggregationUpdateService
