+++ src/main/groovy/de/iteratec/osm/measurement/environment/wptserver/WptResultXml.groovy
@@ -1,4 +1,4 @@
-package de.iteratec.osm.measurement.environment.wptserverproxy
+package de.iteratec.osm.measurement.environment.wptserver
 
 import de.iteratec.osm.measurement.schedule.Job
 import de.iteratec.osm.result.CachedView
