+++ src/test/groovy/de/iteratec/osm/measurement/environment/wptserver/HttpRequestServiceSpec.groovy
@@ -15,7 +15,7 @@
 * limitations under the License.
 */
 
-package de.iteratec.osm.measurement.environment.wptserverproxy
+package de.iteratec.osm.measurement.environment.wptserver
 
 import grails.testing.services.ServiceUnitTest
 import spock.lang.Specification
