+++ src/test/groovy/de/iteratec/osm/measurement/environment/wptserver/PersistScreenshotDependentWptMetricsSpec.groovy
@@ -15,7 +15,7 @@
 * limitations under the License.
 */
 
-package de.iteratec.osm.measurement.environment.wptserverproxy
+package de.iteratec.osm.measurement.environment.wptserver
 
 import de.iteratec.osm.csi.Page
 import de.iteratec.osm.measurement.environment.Browser
