+++ src/test/groovy/de/iteratec/osm/measurement/environment/wptserver/ResultXmlWptServer219Spec.groovy
@@ -1,4 +1,4 @@
-package de.iteratec.osm.measurement.environment.wptserverproxy
+package de.iteratec.osm.measurement.environment.wptserver
 
 import de.iteratec.osm.result.CachedView
 import de.iteratec.osm.result.WptXmlResultVersion
