+++ src/test/groovy/de/iteratec/osm/measurement/environment/wptserver/WptInfrastructurePersistenceWithIncomingResultsSpec.groovy
@@ -1,4 +1,4 @@
-package de.iteratec.osm.measurement.environment.wptserverproxy
+package de.iteratec.osm.measurement.environment.wptserver
 
 import de.iteratec.osm.OsmConfigCacheService
 import de.iteratec.osm.csi.*
